requirements.txt
a2a-sdk==0.2.10
requests>=2.32.0
